pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "delv"
version = "0.1.0"
description = "A tree-based exploration tracking TUI tool for DFS-style research"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "textual>=0.47.0",
    "click>=8.1.0",
    "pyperclip>=1.8.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
delv = "delv.main:main"

[tool.hatch.build.targets.wheel]
packages = ["delv"]

